pyproject.toml
[project]
name = "vehicle-insurance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.54",
    "botocore>=1.42.54",
    "certifi>=2026.1.4",
    "dill>=0.4.1",
    "fastapi>=0.131.0",
    "from-root>=1.3.0",
    "imblearn>=0.0",
    "ipykernel>=7.2.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.8",
    "mypy-boto3-s3>=1.42.37",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "plotly>=6.5.2",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "structlog>=25.5.0",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
